Automatically merge Dependabot PRs to save toil.
We would never not want to merge a security patch from Dependabot in
this repo if the tests are passing. Having a human in the loop is a
waste of time of money as well as an unnecessary security risk.

This is straight from
https://docs.github.com/en/code-security/dependabot/working-with-dependabot/automating-dependabot-with-github-actions#enable-auto-merge-on-a-pull-request
but with the latest version of dependabot/fetch-metadata and without
filtering on a specific dependency.