Remove empty "Add your app to the platform" pages.
Having these prominent links to empty pages just wastes the reader's
time.

We have no plans to write this content any time soon, since we expect
the rate of adding new apps to be on the order of one per year and
therefore any documentation we write now would be outdated by the time
it's used (if it's even discovered at all) and therefore likely to be
counterproductive.