Remove unnecessary use of the word technical.
This repo is in fact the less technical of the two places where GOV.UK
k8s platform documentation can be found. (The other being the config
repos themselves.)

This also makes the subtitle in the header fit on a single line instead
of wrapping and encroaching on the "ALPHA" label.