Rework section on cluster problems for accuracy and relevance.
This is the user documentation and - generally speaking - users aren't
expected to fix problems with the cluster, so let's not try to document
how to troubleshooting a Kubernetes cluster here.

Clarify the intent of this section so that it gives a sense of what a
cluster issue (as opposed to an app issue) might look like and what to
do if they suspect one (i.e. contact Replatforming team, for now).