Remove dependabot_auto_merge workflow.
This doesn't solve the problem that we wanted it to solve (i.e.
eliminating Dependabot toil for this repo) and it isn't likely to in
future. Since it's triggering a bug in the GitHub Slack integration
which causes it to misleadingly say "0/1 checks passed" on all
non-Dependabot PRs, let's drop it.