Add kubectl tutorial to walk through some common developer tasks
This follows the same steps the Platform Reliability team took in
an internal workshop ([1]). The workshop was well received and
the team felt more confident interacting with the new platform
(at a low level) once they'd had a guided introduction to it, which
was missing from the other two tutorials.

[1]: https://docs.google.com/document/d/17ncZBw2gcafk37mIEFtrQi1DME1uBxGjyHNPHM1SVgo/edit