Update team name and consolidate contact details.
Update the team name and contact details to reflect the recent name
change. Put the team contact details on a single page (along with a bit
of context about the sort of thing we do) so it might be slightly less
hassle next time.

Also update a few outdated odds and ends encountered along the way and
stick a health hazard warning on the cluster turnup instructions for the
time being (lest they lead some poor innocent soul astray).